Added Generic Math to BitUtils (#3929)
* Generic Math Update

Updated Several functions in Ryujinx.Common/Utilities/BitUtils to use generic math

* Updated BitUtil calls

* Removed Whitespace

* Switched decrement

* Fixed changed method calls.

The method calls were originally changed on accident due to me relying too much on intellisense doing stuff for me

* Update Ryujinx.Common/Utilities/BitUtils.cs

+++ Ryujinx.HLE/HOS/Kernel/Memory/KPageHeap.cs
@@ -48,7 +48,7 @@ namespace Ryujinx.HLE.HOS.Kernel.Memory
                 {
                     int diff = 1 << (NextShift - Shift);
 
-                    offset = BitUtils.AlignDown(offset, diff);
+                    offset = BitUtils.AlignDown(offset, (ulong)diff);
 
                     if (_bitmap.ClearRange(offset, diff))
                     {
